+++ Examples/SearchExamples/MapboxMapsCategoryResultsViewController.swift
@@ -3,6 +3,7 @@ import MapboxSearch
 
 class MapboxMapsCategoryResultsViewController: MapsViewController {
     let searchEngine = CategorySearchEngine()
+//    let searchEngine = CategorySearchEngine(accessToken: "<#You can pass access token manually#>")
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)