+++ Sources/MapboxSearch/PublicAPI/ServiceProvider.swift
@@ -1,7 +1,6 @@
 import Foundation
 
 let accessTokenPlistKey = "MBXAccessToken"
-let legacyAccessTokenPlistKey = "MGLMapboxAccessToken"
 let baseURLPlistKey = "MapboxAPIBaseURL"
 
 protocol ServiceProviderProtocol {
@@ -15,6 +14,7 @@ protocol ServiceProviderProtocol {
 protocol EngineProviderProtocol {
     func createEngine(
         apiType: CoreSearchEngine.ApiType,
+        accessToken: String,
         locationProvider: CoreLocationProvider?
     ) -> CoreSearchEngineProtocol
 
@@ -53,17 +53,20 @@ public class ServiceProvider: ServiceProviderProtocol {
 extension ServiceProvider: EngineProviderProtocol {
     func getStoredAccessToken() -> String? {
         Bundle.main.object(forInfoDictionaryKey: accessTokenPlistKey) as? String
-        ?? Bundle.main.object(forInfoDictionaryKey: legacyAccessTokenPlistKey) as? String
     }
+
 
     func createEngine(
         apiType: CoreSearchEngine.ApiType,
+        accessToken: String,
         locationProvider: CoreLocationProvider?
     ) -> CoreSearchEngineProtocol {
         // UserDefaults can be used to setup base url in runtime (e.g. UI tests)
         // UserDefaults can be used to setup base url in runtime (e.g. UI tests)
         let defaultsBaseURL = UserDefaults.standard.value(forKey: baseURLPlistKey) as? String
         let bundleBaseURL = Bundle.main.object(forInfoDictionaryKey: baseURLPlistKey) as? String
+
+        MapboxOptions.accessToken = accessToken
 
         let engineOptions = CoreSearchEngine.Options(
             baseUrl: bundleBaseURL ?? defaultsBaseURL,