+++ Tests/MapboxSearchIntegrationTests/CategorySearchEngineIntegrationTests.swift
@@ -5,6 +5,7 @@ import CoreLocation
 class CategorySearchEngineIntegrationTests: MockServerTestCase {
     
     lazy var searchEngine = CategorySearchEngine(
+        accessToken: "access-token",
         locationProvider: DefaultLocationProvider(),
         supportSBS: true
     )