+++ Tests/MapboxSearchIntegrationTests/PlaceAutocompleteIntegrationTests.swift
@@ -11,6 +11,7 @@ final class PlaceAutocompleteIntegrationTests: MockServerTestCase {
         super.setUp()
 
         placeAutocomplete = PlaceAutocomplete(
+            accessToken: "access-token",
             locationProvider: DefaultLocationProvider()
         )
     }