+++ Tests/MapboxSearchTests/Common/Stubs&Models/ServiceProviderStub.swift
@@ -14,9 +14,9 @@ class ServiceProviderStub: ServiceProviderProtocol, EngineProviderProtocol {
 
     var latestCoreEngine: CoreSearchEngineStub!
 
-    func createEngine(apiType: CoreSearchEngine.ApiType, locationProvider: CoreLocationProvider?) -> CoreSearchEngineProtocol {
+    func createEngine(apiType: CoreSearchEngine.ApiType, accessToken: String, locationProvider: CoreLocationProvider?) -> CoreSearchEngineProtocol {
         let locationProvider: CoreLocationProvider? = WrapperLocationProvider(wrapping: DefaultLocationProvider())
-        latestCoreEngine = CoreSearchEngineStub(location: locationProvider)
+        latestCoreEngine = CoreSearchEngineStub(accessToken: "mapbox-access-token", location: locationProvider)
         return latestCoreEngine
     }
 }