+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Guide: https://keepachangelog.com/en/1.0.0/
 
 <!-- Add changes for active work here -->
 
+- [Core] Remove legacy `MGLMapboxAccessToken`.
+
 ## 2.0.0-alpha.1
 
 ### Breaking changes
@@ -19,8 +21,6 @@ Guide: https://keepachangelog.com/en/1.0.0/
 - [Place Autocomplete] Suggestions no longer perform a `retrieve` call.
 - [Place Autocomplete] `PlaceAutocomplete.Suggestion` field `coordinate: CLLocationCoordinate2D?` is now an optional.
 - [Place Autocomplete] `Result.coordinate` is now an optional.
-- [Core] Updated to Xcode 14.1 minimum version
-- [Core] Updated deployment target to iOS 12
 - [Core] Remove bitcode support
 - [Core] Updated API usage:
 	- Removed parameter-based Access Token. Be sure to provide your token in Info.plist.